+++ api.py
@@ -2,8 +2,7 @@ from fastapi import FastAPI, HTTPException
 from fastapi.staticfiles import StaticFiles
 from fastapi.middleware.cors import CORSMiddleware
 from mydates import getListaFeriados
-from myClasses import IntegracaoEscala
-from myClientSupabase import get_credentials
+from myConfigSupabase import getConfig
 
 app = FastAPI()
 
@@ -20,13 +19,13 @@ async def get_lista_feriados(ano: int):
     lista = getListaFeriados(ano)
 
     if not lista:
-        raise HTTPException(status_code=404, detail="Erro ao buscar lista de feriados")
+        raise HTTPException(status_code=503, detail="Erro ao buscar lista de feriados")
 
     return {"listaFeriados": lista}
 
-@app.get("/integracaoSupabase")
-async def get_integracao_escala():
-    return get_credentials()
+@app.get("/configSupabase")
+async def get_config_supabase():
+    return getConfig()
 
 static = FastAPI()
