+++ requirements.txt
@@ -1,5 +1,4 @@
 fastapi==0.115.12
 uvicorn==0.34.2
 workalendar==17.0.0
-supabase==2.15.1
 python-dotenv==1.1.0
